Alpha

+++ userge/plugins/fun/kang.py
@@ -8,33 +8,36 @@
 #
 # All rights reserved.
 
+import io
 import os
 import random
-import urllib.request
+import emoji
 
 from PIL import Image
-from pyrogram.api.functions.messages import GetStickerSet
-from pyrogram.api.types import InputStickerSetShortName
-from pyrogram.errors.exceptions.bad_request_400 import YouBlockedUser
+from pyrogram.raw.functions.messages import GetStickerSet
+from pyrogram.raw.types import InputStickerSetShortName
+from pyrogram.errors import YouBlockedUser, StickersetInvalid
 
-from userge import userge, Message, Config, pool
+from userge import userge, Message, Config
 
 
 @userge.on_cmd(
     "kang", about={
         'header': "kangs stickers or creates new ones",
+        'flags': {
+            '-s': "without link",
+            '-d': "without trace"},
         'usage': "Reply {tr}kang [emoji('s)] [pack number] to a sticker or "
                  "an image to kang it to your userbot pack.",
-        'examples': ["{tr}kang", "{tr}kang 🤔", "{tr}kang 2", "{tr}kang 🤔 2"]},
+        'examples': ["{tr}kang", "{tr}kang -s", "{tr}kang -d",
+                     "{tr}kang 🤔", "{tr}kang 2", "{tr}kang 🤔 2"]},
     allow_channels=False, allow_via_bot=False)
 async def kang_(message: Message):
     """ kang a sticker """
     user = await userge.get_me()
-    if not user.username:
-        user.username = user.first_name or user.id
     replied = message.reply_to_message
     photo = None
-    emoji = None
+    emoji_ = None
     is_anim = False
     resize = False
     if replied and replied.media:
@@ -45,7 +48,10 @@ async def kang_(message: Message):
         elif replied.document and "tgsticker" in replied.document.mime_type:
             is_anim = True
         elif replied.sticker:
-            emoji = replied.sticker.emoji
+            if not replied.sticker.file_name:
+                await message.edit("`Sticker has no Name!`")
+                return
+            emoji_ = replied.sticker.emoji
             is_anim = replied.sticker.is_animated
             if not replied.sticker.file_name.endswith('.tgs'):
                 resize = True
@@ -59,19 +65,29 @@ async def kang_(message: Message):
         await message.edit("`I can't kang that...`")
         return
     if photo:
-        args = message.input_str.split()
-        if not emoji:
-            emoji = "🤔"
+        args = message.filtered_input_str.split()
         pack = 1
         if len(args) == 2:
-            emoji, pack = args
+            emoji_, pack = args
         elif len(args) == 1:
             if args[0].isnumeric():
                 pack = int(args[0])
             else:
-                emoji = args[0]
-        packname = f"a{user.id}_by_{user.username}_{pack}"
-        packnick = f"@{user.username}'s kang pack Vol.{pack}"
+                emoji_ = args[0]
+
+        if emoji_ and emoji_ not in emoji.UNICODE_EMOJI:
+            emoji_ = None
+        if not emoji_:
+            emoji_ = "🤔"
+
+        u_name = user.username
+        if u_name:
+            u_name = "@" + u_name
+        else:
+            u_name = user.first_name or user.id
+        packname = f"a{user.id}_by_userge_{pack}"
+        custom_packnick = Config.CUSTOM_PACK_NAME or f"{u_name}'s kang pack"
+        packnick = f"{custom_packnick} Vol.{pack}"
         cmd = '/newpack'
         if resize:
             photo = resize_photo(photo)
@@ -79,15 +95,15 @@ async def kang_(message: Message):
             packname += "_anim"
             packnick += " (Animated)"
             cmd = '/newanimated'
-
-        @pool.run_in_thread
-        def get_response():
-            response = urllib.request.urlopen(  # nosec
-                urllib.request.Request(f'http://t.me/addstickers/{packname}'))
-            return response.read().decode("utf8").split('\n')
-        htmlstr = await get_response()
-        if ("  A <strong>Telegram</strong> user has created "
-                "the <strong>Sticker&nbsp;Set</strong>.") not in htmlstr:
+        exist = False
+        try:
+            exist = await message.client.send(
+                GetStickerSet(
+                    stickerset=InputStickerSetShortName(
+                        short_name=packname)))
+        except StickersetInvalid:
+            pass
+        if exist is not False:
             async with userge.conversation('Stickers', limit=30) as conv:
                 try:
                     await conv.send_message('/addsticker')
@@ -100,8 +116,11 @@ async def kang_(message: Message):
                 limit = "50" if is_anim else "120"
                 while limit in msg.text:
                     pack += 1
-                    packname = f"a{user.id}_by_{user.username}_{pack}"
-                    packnick = f"@{user.username}'s kang pack Vol.{pack}"
+                    packname = f"a{user.id}_by_userge_{pack}"
+                    packnick = f"{custom_packnick} Vol.{pack}"
+                    if is_anim:
+                        packname += "_anim"
+                        packnick += " (Animated)"
                     await message.edit("`Switching to Pack " + str(pack) +
                                        " due to insufficient space`")
                     await conv.send_message(packname)
@@ -113,21 +132,23 @@ async def kang_(message: Message):
                         await conv.get_response(mark_read=True)
                         await conv.send_document(photo)
                         await conv.get_response(mark_read=True)
-                        await conv.send_message(emoji)
+                        await conv.send_message(emoji_)
                         await conv.get_response(mark_read=True)
                         await conv.send_message("/publish")
                         if is_anim:
                             await conv.get_response(mark_read=True)
-                            await conv.send_message(f"<{packnick}>")
+                            await conv.send_message(f"<{packnick}>", parse_mode=None)
                         await conv.get_response(mark_read=True)
                         await conv.send_message("/skip")
                         await conv.get_response(mark_read=True)
                         await conv.send_message(packname)
                         await conv.get_response(mark_read=True)
-                        await message.edit(
-                            f"`Sticker added in a Different Pack !\n"
-                            "This Pack is Newly created!\n"
-                            f"Your pack can be found [here](t.me/addstickers/{packname})")
+                        if '-d' in message.flags:
+                            await message.delete()
+                        else:
+                            out = "__kanged__" if '-s' in message.flags else \
+                                f"[kanged](t.me/addstickers/{packname})"
+                            await message.edit(f"**Sticker** {out} __in a Different Pack__**!**")
                         return
                 await conv.send_document(photo)
                 rsp = await conv.get_response(mark_read=True)
@@ -135,7 +156,7 @@ async def kang_(message: Message):
                     await message.edit("`Failed to add sticker, use` @Stickers "
                                        "`bot to add the sticker manually.`")
                     return
-                await conv.send_message(emoji)
+                await conv.send_message(emoji_)
                 await conv.get_response(mark_read=True)
                 await conv.send_message('/done')
                 await conv.get_response(mark_read=True)
@@ -153,23 +174,28 @@ async def kang_(message: Message):
                 await conv.send_document(photo)
                 rsp = await conv.get_response(mark_read=True)
                 if "Sorry, the file type is invalid." in rsp.text:
-                    await args.edit("`Failed to add sticker, use` @Stickers "
-                                    "`bot to add the sticker manually.`")
+                    await message.edit("`Failed to add sticker, use` @Stickers "
+                                       "`bot to add the sticker manually.`")
                     return
-                await conv.send_message(emoji)
+                await conv.send_message(emoji_)
                 await conv.get_response(mark_read=True)
                 await conv.send_message("/publish")
                 if is_anim:
                     await conv.get_response(mark_read=True)
-                    await conv.send_message(f"<{packnick}>")
+                    await conv.send_message(f"<{packnick}>", parse_mode=None)
                 await conv.get_response(mark_read=True)
                 await conv.send_message("/skip")
                 await conv.get_response(mark_read=True)
                 await conv.send_message(packname)
                 await conv.get_response(mark_read=True)
-        await message.edit(f"`Sticker kanged successfully!`\n"
-                           f"Pack can be found [here](t.me/addstickers/{packname})")
-        os.remove(photo)
+        if '-d' in message.flags:
+            await message.delete()
+        else:
+            out = "__kanged__" if '-s' in message.flags else \
+                f"[kanged](t.me/addstickers/{packname})"
+            await message.edit(f"**Sticker** {out}**!**")
+        if os.path.exists(str(photo)):
+            os.remove(photo)
 
 
 @userge.on_cmd("stkrinfo", about={
@@ -204,22 +230,18 @@ async def sticker_pack_info_(message: Message):
     await message.edit(out_str)
 
 
-def resize_photo(photo: str) -> str:
+def resize_photo(photo: str) -> io.BytesIO:
     """ Resize the given photo to 512x512 """
     image = Image.open(photo)
-    maxsize = (512, 512)
-    if max(image.width, image.height) < 512:
-        scale = 512 / max(image.width, image.height)
-        new_size = (int(image.width * scale), int(image.height * scale))
-        image = image.resize(new_size, Image.ANTIALIAS)
-    else:
-        image.thumbnail(maxsize)
+    maxsize = 512
+    scale = maxsize / max(image.width, image.height)
+    new_size = (int(image.width*scale), int(image.height*scale))
+    image = image.resize(new_size, Image.LANCZOS)
+    resized_photo = io.BytesIO()
+    resized_photo.name = "sticker.png"
+    image.save(resized_photo, "PNG")
     os.remove(photo)
-    photo = os.path.join(Config.DOWN_PATH, "sticker.png")
-    if os.path.exists(photo):
-        os.remove(photo)
-    image.save(photo, "PNG")
-    return photo
+    return resized_photo
 
 
 KANGING_STR = (
